fix: use unix-style paths for action invocation
Fixes and closes Caphyon/advinst-github-action#9, which describes
a failure on macOS or Linux runners when the action is otherwise
inert.

Github runners accept Unix-style paths always, whereas Windows
style paths seem to fail on non-Windows runners. Thus, this change
should be transparent for Windows users and should fix the error
filed in #9 for macOS and Linux runners.

Signed-off-by: Sam Gammon <sam@elide.ventures>